fix(db): use INT for dog_trial_jump_heights.jump_height_inches
Elected jump height is an AKC-defined integer bucket set (Obedience:
4, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32, 34, 36; Rally:
4, 8, 12, 16). The CHECK already enumerated those 15 integer values;
storing them as NUMERIC(4,1) bought nothing and cost storage.
Inserting 12.0 would succeed but print as "12.0" from a column whose
domain is strictly integers, which is semantic opacity.

This column is semantically distinct from dogs.jump_height_measured,
which is correctly NUMERIC(4,1) because it records a physical
measurement at the withers (13.5" is real). The two columns share a
name root but not a type or meaning: elected bucket vs measured
value.

The original PR 2b rationale ("match jump_heights.height_inches from
PR 2a") conflated the two. jump_heights (reference catalog) stays
NUMERIC(4,1) from PR 2a; if AKC ever adds half-inch elected buckets
that deserves its own design conversation and would need a matching
CHECK rewrite, neither of which is happening now.

Landed as a new commit rather than amending 1cdcd3a so the decision
trail is visible in the branch history: the original choice, the
investigation that caught it, and the correction are all observable.

Changes:
 * Column type: NUMERIC(4, 1) -> INT
 * CHECK constraint: unchanged (same 15 integer values)
 * .down.sql: unchanged (DROP TABLE is type-agnostic)
 * Header comment updated to reflect the new rationale

Round-trip verified: 42 up -> 35 after PR 2b revert -> 44 after re-up.
CHECK still rejects non-bucket values (13 raises check_violation).

+++ db/migrations/20260424120200_create_dog_trial_jump_heights.up.sql
@@ -9,11 +9,15 @@
 -- lives on one row per (dog, trial) rather than per entry line.
 -- Rally Choice entries do not consult this table.
 --
--- jump_height_inches is NUMERIC(4,1) to match
--- jump_heights.height_inches from PR 2a. All 2026 seed values are
--- whole inches; keeping NUMERIC leaves room for future half-inch
--- heights without a later ALTER and avoids a cross-type cast when
--- joining against jump_heights or a future FK validation.
+-- jump_height_inches is INT. Elected jump height is an AKC-defined
+-- integer bucket set (Obedience: 4, 8, 10, 12, 14, 16, 18, 20, 22,
+-- 24, 26, 28, 30, 32, 34, 36; Rally: 4, 8, 12, 16). This is
+-- semantically distinct from dogs.jump_height_measured NUMERIC(4,1),
+-- which is a physical measurement at the withers and can be
+-- fractional (13.5"). The two columns share a name root but not a
+-- type or meaning. If AKC ever adds half-inch elected buckets, that
+-- change deserves its own design conversation; until then, INT plus
+-- the enumerated CHECK is the honest shape.
 --
 -- The CHECK constraint enumerates the legitimate AKC heights.
 -- On-leash classes do not consult this table at all (they have no
@@ -41,7 +45,7 @@ CREATE TABLE dog_trial_jump_heights (
     -- jump-height rows go with it; they are meaningless outside
     -- that trial.
     trial_id                        UUID NOT NULL REFERENCES trials(id) ON DELETE CASCADE,
-    jump_height_inches              NUMERIC(4, 1) NOT NULL,
+    jump_height_inches              INT NOT NULL,
     -- TRUE when a judge doubted the submitted height and measured
     -- the dog in-ring. The override applies to every remaining
     -- entry for this dog at this trial; the app layer UPDATEs the
